+++ src/Pattern/Match.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phinder;
+namespace Phinder\Pattern;
 
 final class Match
 {
